Fixed with roles and Auth

+++ app/Http/Controllers/Admin/SchoolManagementController.php
@@ -4,14 +4,24 @@ namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
 use App\Models\SchoolManagement;
+use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Hash;
 
 class SchoolManagementController extends Controller
 {
     /**
      * Display a listing of the resource.
-     */
+      */
+      protected $AuthUser;
+
+        public function __construct()
+        {
+            $this->AuthUser=Auth::user();
+        }
+
     public function index()
     {
         $data=SchoolManagement::all();
@@ -24,6 +34,9 @@ class SchoolManagementController extends Controller
      */
     public function create()
     {
+        if(!$this->AuthUser->hasRole(1))
+         return redirect()->back();
+
         $title='School Create';
         return view('admin.pages.school_managements.create',compact('title'));
     }
@@ -35,6 +48,10 @@ class SchoolManagementController extends Controller
     {
          $request->validate([
          'name'=>'required|string',
+         'email'=>'required|email|unique:school_management,email,'.$request->id.',id,deleted_at,NULL|unique:users,email,'.$request->user_id.',id,deleted_at,NULL',
+         'phone'=>'required|numeric|digits_between:10,12|unique:school_management,phone,'.$request->id.',id,deleted_at,NULL|unique:users,mobile,'.$request->user_id.',id,deleted_at,NULL',
+         'password'=>'nullable|min:8',
+         'confirm_password'=>'required_with:password|same:password',
          'address'=>'required|string',
          'open_time'=>'required',
          'close_time'=>'required',
@@ -82,9 +99,25 @@ class SchoolManagementController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, string $id)
+    public function update(Request $request, SchoolManagement $schoolManagement)
     {
-        //
+        $request->validate(
+            [
+                'newpassword'=>'required|min:8',
+                'confirm_password'=>'required_with:newpassword|same:newpassword',
+            ]
+            );
+         try
+         {
+        User::where('id',$schoolManagement->user_id)->update(['password'=>Hash::make($request->newpassword)]);
+
+        return redirect()->route('teachers.index')->with('change','Password Change Successfully');
+
+         }
+         catch(\Exception $e)
+         {
+            $e->getMessage();
+         }
     }
 
     /**
@@ -92,6 +125,11 @@ class SchoolManagementController extends Controller
      */
     public function destroy(SchoolManagement $schoolManagement)
     {
+        if(!$this->AuthUser->hasRole(1))
+         return redirect()->back();
+
+        $user=User::find($schoolManagement->user_id);
+        $user->delete();
         $schoolManagement->delete();
         return redirect()->back();
     }
